chore: cut monitor scope back to merge lane
