+++ frontend/monitor/index.html
@@ -3,7 +3,6 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <title>Leon Sandbox Monitor</title>
   </head>
   <body>
@@ -11,3 +10,4 @@
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
+
