+++ frontend/monitor/src/main.tsx
@@ -4,5 +4,8 @@ import App from "./App";
 import "./styles.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
-  <App />,
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
 );
+
